+++ cider-util.el
@@ -107,6 +107,10 @@ If BUFFER is provided act on that buffer instead."
   "Return the starting position of the current defun."
   (car (cider--region-for-defun-at-point)))
 
+(defun cider-defun-at-point-end-pos ()
+  "Return the end position of the current defun."
+  (cadr (cider--region-for-defun-at-point)))
+
 (defun cider-ns-form ()
   "Retrieve the ns form."
   (when (clojure-find-ns)